added ip checking for anything other than search

+++ routes/recipe.js
@@ -4,13 +4,16 @@ const recipe = (app,DB,creds) => {
 
 	/********************************************/
 	app.delete('/api/recipe/:id', (request, response) => {
-		const Recipe = require('../classes/recipe');
-		let recipe = new Recipe(DB);
+		const remoteAddress = request.headers['x-forwarded-for'] || request.connection.remoteAddress;
+		if(remoteAddress.includes("192.168.1.")){
+			const Recipe = require('../classes/recipe');
+			let recipe = new Recipe(DB);
 
-		recipe.id = request.params.id;
-		recipe.delete();
-		
-		response.json({'status': 'ok'});
+			recipe.id = request.params.id;
+			recipe.delete();
+			
+			response.json({'status': 'ok'});
+		}
 	});
 	/********************************************/
 
@@ -30,60 +33,72 @@ const recipe = (app,DB,creds) => {
 
 	/********************************************/
 	app.post('/api/recipe', (request, response) => {
-		const Recipe = require('../classes/recipe');
-		
-		let recipe = new Recipe(DB);
-		recipe.post = request.body;
-		let id = recipe.add();
+		const remoteAddress = request.headers['x-forwarded-for'] || request.connection.remoteAddress;
+		if(remoteAddress.includes("192.168.1.")){
+			const Recipe = require('../classes/recipe');
+			
+			let recipe = new Recipe(DB);
+			recipe.post = request.body;
+			let id = recipe.add();
 
-		let out = {
-			id: id
-		};
-		response.json(out);
+			let out = {
+				id: id
+			};
+			response.json(out);
+		}
 	});
 	/********************************************/
 
 
 	/********************************************/
 	app.patch('/api/recipe/:id', (request, response) => {
-		const Recipe = require('../classes/recipe');
-		
-		let recipe = new Recipe(DB);
-		recipe.post = request.body;
-		recipe.id = request.params.id;
-		recipe.update();
+		const remoteAddress = request.headers['x-forwarded-for'] || request.connection.remoteAddress;
+		if(remoteAddress.includes("192.168.1.")){
+			const Recipe = require('../classes/recipe');
+			
+			let recipe = new Recipe(DB);
+			recipe.post = request.body;
+			recipe.id = request.params.id;
+			recipe.update();
 
-		response.json({'status': 'ok'});
+			response.json({'status': 'ok'});
+		}
 	});
 	/********************************************/
 
 
 	/********************************************/
 	app.post('/api/recipe/:id/sendgrid', (request, response) => {
-		const Recipe = require('../classes/recipe');
-		
-		let recipe = new Recipe(DB);
-		recipe.creds = creds;
-		recipe.id = request.params.id;
-		recipe.post = request.body;
-		recipe.sendgrid();
+		const remoteAddress = request.headers['x-forwarded-for'] || request.connection.remoteAddress;
+		if(remoteAddress.includes("192.168.1.")){
+			const Recipe = require('../classes/recipe');
+			
+			let recipe = new Recipe(DB);
+			recipe.creds = creds;
+			recipe.id = request.params.id;
+			recipe.post = request.body;
+			recipe.sendgrid();
 
-		response.json({'status': 'sent'});
+			response.json({'status': 'sent'});
+		}
 	});
 	/********************************************/
 
 
 	/********************************************/
 	app.post('/api/recipe/:id/gmail', (request, response) => {
-		const Recipe = require('../classes/recipe');
-		
-		let recipe = new Recipe(DB);
-		recipe.creds = creds;
-		recipe.id = request.params.id;
-		recipe.post = request.body;
-		recipe.gmail();
+		const remoteAddress = request.headers['x-forwarded-for'] || request.connection.remoteAddress;
+		if(remoteAddress.includes("192.168.1.")){
+			const Recipe = require('../classes/recipe');
+			
+			let recipe = new Recipe(DB);
+			recipe.creds = creds;
+			recipe.id = request.params.id;
+			recipe.post = request.body;
+			recipe.gmail();
 
-		response.json({'status': 'sent'});
+			response.json({'status': 'sent'});
+		}
 	});
 	/********************************************/
